AcrylixOld
An older version of Acrylix (1.4). It will not be updated unfortunately.
Many people (mainly Arceus X users) were complaining that the new version does not work on their executor, so I thought maybe upload the old version so they will be able to use it.
The new version is expected to load faster than the old one, they both depend on the speed of your network connection.